fix: update meeting page PageHead
