+++ packages/web/src/features/meeting/types/meeting.ts
@@ -27,6 +27,19 @@ export const getMeetingEnumFromValue = (value: string | null) => {
   }
 };
 
+export const meetingEnumToText = (meetingEnum: string) => {
+  switch (meetingEnum) {
+    case MeetingEnum.clubRepresentativesCouncilMeeting.toString():
+      return "전체동아리대표자회의";
+    case MeetingEnum.expansiveOperativeCommittee.toString():
+      return "확대운영위원회";
+    case MeetingEnum.operativeCommittee.toString():
+      return "운영위원회";
+    default:
+      return "분과회의";
+  }
+};
+
 export interface MeetingNoticeItemType {
   id: number;
   meetingEnumId: MeetingEnum;
